+++ dspace-api/src/main/java/org/dspace/importer/external/SourceExceptionHandler.java
@@ -8,13 +8,13 @@
 
 package org.dspace.importer.external;
 
-import org.dspace.importer.external.service.other.MetadataSource;
+import org.dspace.importer.external.service.other.AbstractRemoteMetadataSource;
 
 /**
  * Represent a  handler that forces implementations to define their own behaviour for exceptions originating from
  * @author: Antoine Snyers (antoine at atmire dot com)
  */
-public abstract interface SourceExceptionHandler<T extends MetadataSource> {
+public abstract interface SourceExceptionHandler<T extends AbstractRemoteMetadataSource> {
 
     /**
      * Represents a method contract to handle Exceptions originating from the source in a specific way
