+++ dspace-api/src/main/java/org/dspace/importer/external/pubmed/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Pubmed implementation of {@link org.dspace.importer.external.service.other.Imports}
+ * Pubmed implementation of {@link org.dspace.importer.external.service.other.MetadataSource}
  * Based on metadata support in {@link org.dspace.importer.external.metadatamapping}
  * @author Roeland Dillen (roeland at atmire dot com)
  */
